requirements.txt
# Requirements for Raspberry Pi 5 (Python 3.12+)
# Updated versions compatible with Python 3.12

# Core Web Framework
Flask==3.0.0
Flask-Login==0.6.3
Flask-CORS==4.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy>=2.0.35
PyMySQL==1.1.0
cryptography==41.0.7

# Password Hashing
bcrypt==4.1.2

# System Monitoring
psutil==5.9.6

# Computer Vision & AI - Updated for Python 3.12
opencv-python==4.10.0.84
ultralytics>=8.3.0
torch>=2.5.0
torchvision>=0.20.0
numpy>=1.26.0
Pillow>=11.0.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0
scipy>=1.11.0

# MQTT (for IoT communication)
paho-mqtt==1.6.1

# Utilities
python-dotenv==1.0.0
